+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and added an optional 3rd argument `(bool)` if the script should be interrupted
 - added function to throw an custom Error
 - added `startmenu` and `local_startmenu` from `some_default_dirs` to
 `dapi_io.get_default_directories`
+- added clear - a function to clear the Terminal Window Content
 
 ## 0.1.13 - 10.10.2025
 - add auto publish for the mkdocs documentation